pyproject.toml
[tool.poetry]
name = "photo_editor"
version = "0.1.0"
description = ""
authors = ["Герман Гриценко <gerfest.gricenko@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-decouple = "^3.6"
celery = "^5.2.7"
rabbitmq = "^0.2.0"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
